+++ R/tree_accessors.R
@@ -406,8 +406,6 @@ setMethod(
 )
 
 
-
-
 #' @rdname int_methods
 setGeneric("pos_splvals", function(obj) standardGeneric("pos_splvals"))
 
@@ -2864,7 +2862,6 @@ setMethod(
 )
 
 
-
 #' @rdname int_methods
 #' @export
 setGeneric(
@@ -4124,8 +4121,6 @@ setMethod("section_div", "TableRow", function(obj) {
 })
 
 
-
-
 # section_div setter from table object
 #' @rdname section_div
 #' @export
@@ -4621,3 +4616,73 @@ setMethod(
   "obj_stat_names", "RowsVerticalSection",
   function(obj) lapply(obj, obj_stat_names)
 )
+
+# obj_round_type getter ---------------------------------------------------------------
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type
+setMethod("obj_round_type", "ANY", function(obj) attr(obj, "round_type", exact = TRUE))
+
+#' generics and methods for obj_round_type
+#'
+#' These are internal methods that are documented only to satisfy `R CMD check`. End users should pay no
+#' attention to this documentation, except for the few exported methods.
+#'
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type
+#' @export
+setMethod("obj_round_type", "PreDataTableLayouts", function(obj) obj@round_type)
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type
+#' @export
+setMethod("obj_round_type", "VTableTree", function(obj) obj@round_type)
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type
+setMethod("obj_round_type", "TableRow", function(obj) obj@round_type)
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type
+setMethod("obj_round_type", "CellValue", function(obj) attr(obj, "round_type", exact = TRUE))
+
+# obj_round_type setter ---------------------------------------------------------------
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type<-
+#' @export
+setMethod("obj_round_type<-", "VTableTree", function(obj, value) {
+  stopifnot(length(value) == 1)
+  checkmate::assert_subset(value, valid_round_type)
+  obj@round_type <- value
+  # also set round_type to slot in children class
+  tree_children(obj) <- lapply(tree_children(obj), `obj_round_type<-`, value = value)
+  if (!is.null(content_table(obj))) {
+    ctab <- content_table(obj)
+    obj_round_type(ctab) <- value
+  }
+  obj
+})
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type<-
+# not useful for end user, needed for recursive approach
+setMethod("obj_round_type<-", "TableRow", function(obj, value) {
+  obj@round_type <- value
+  row_cells(obj) <- lapply(row_cells(obj), `obj_round_type<-`, value = value)
+  obj
+})
+
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type<-
+# not useful for end user, needed for recursive approach
+setMethod("obj_round_type<-", "LabelRow", function(obj, value) {
+  obj
+})
+
+
+#' @rdname formatters_methods
+#' @exportMethod obj_round_type<-
+setMethod("obj_round_type<-", "CellValue", function(obj, value) {
+  attr(obj, "round_type") <- value
+  obj
+})